Fix hard coded year in summaries
Fix hard coded year in summaries, which was always set to 2021.  This
may not always be correct however, since summaries can sometimes be
for the previous year if invoked in January.  Likely a change server
side should be made to return the time period for the summary along
with the data.